entity files have been renamed
