+++ server/entity/serializable.py
